backend/app/requirements.txt
paho-mqtt==2.1.0
sentinel==1.0.0
